+++ .github/workflows/tests.yml
@@ -47,7 +47,7 @@ jobs:
     - name: Test Notebooks
       run: |
         . ./venv/bin/activate
-        pytest --nbmake notebooks -n=auto --nbmake-kernel=python3 --nbmake-timeout=600  # 10 minutes timeout
+        pytest --nbmake tutorials -n=auto --nbmake-kernel=python3 --nbmake-timeout=600  # 10 minutes timeout
 
     - name: Test Unittests with pytest
       run: |
